+++ starter-kits/nextjs-starter-kit/app/text/[id]/TextDocumentView.tsx
@@ -1,14 +1,11 @@
 "use client";
 
-import { LiveMap } from "@liveblocks/client";
-import { RoomProvider } from "@liveblocks/react/suspense";
 import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { DocumentHeader, DocumentHeaderSkeleton } from "@/components/Document";
 import { TextEditor } from "@/components/TextEditor";
-import { DocumentLayout } from "@/layouts/Document";
+import { DocumentLayout, DocumentProviders } from "@/layouts/Document";
 import { ErrorLayout } from "@/layouts/Error";
-import { InitialDocumentProvider } from "@/lib/hooks";
 import { Document, ErrorData } from "@/types";
 
 type Props = {
@@ -36,18 +33,12 @@ export function TextDocumentView({ initialDocument, initialError }: Props) {
   }
 
   return (
-    <RoomProvider
-      id={id as string}
-      initialPresence={{ cursor: null }}
-      initialStorage={{ notes: new LiveMap() }}
-    >
-      <InitialDocumentProvider initialDocument={initialDocument}>
-        <DocumentLayout
-          header={<DocumentHeader documentId={initialDocument.id} />}
-        >
-          <TextEditor />
-        </DocumentLayout>
-      </InitialDocumentProvider>
-    </RoomProvider>
+    <DocumentProviders roomId={id} initialDocument={initialDocument}>
+      <DocumentLayout
+        header={<DocumentHeader documentId={initialDocument.id} />}
+      >
+        <TextEditor />
+      </DocumentLayout>
+    </DocumentProviders>
   );
 }