+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.11.1 (Yet to be published)
+
+### `@liveblocks/react-lexical`
+
+- Fix an issue with `AnchoredThreads` component not working correctly on certain
+  React versions.
+
+### `@liveblocks/react-tiptap`
+
+- Fix an issue with `AnchoredThreads` component not working correctly on certain
+  React versions.
+
 ## 2.11.0
 
 ### `@liveblocks/react-ui`